関数のコードを1ファイルにまとめる
